use float instead of int to seek

+++ app/containers/Player/index.js
@@ -106,7 +106,7 @@ class Player extends React.Component {
 
   setSeek = seconds =>
     // eslint-disable-next-line radix
-    this.playerRef && this.playerRef.seekTo(parseInt(seconds));
+    this.playerRef && this.playerRef.seekTo(parseFloat(seconds));
 
   handleProgress = played => this.setState({ played: played.playedSeconds });
